fixes for double-prec compile

+++ source/gui/particlepainter.cpp
@@ -101,7 +101,9 @@ static inline void glVertex(const Vec3& v, Real dx) {
 }
 
 static inline void glColor(const Vec3& color) {
-    glColor3f(std::max(0.0f,std::min(1.0f,color.x)), std::max(0.0f,std::min(1.0f,color.y)), std::max(0.0f,std::min(1.0f,color.z)));
+    glColor3f( std::max(0.0f,std::min(1.0f,(float)color.x)), 
+			   std::max(0.0f,std::min(1.0f,(float)color.y)), 
+			   std::max(0.0f,std::min(1.0f,(float)color.z)) );
 }
 
 void ParticlePainter::paint() {
